pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cctyper"
version = "1.8.0"
description = "CRISPRCasTyper: Automatic detection and subtyping of CRISPR-Cas operons"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Jakob Russel"}]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 4 - Beta",
]
dependencies = [
    "numpy >= 1.17.5",
    "pandas >= 1.3",
    "scipy >= 1.4.1",
    "biopython >= 1.76",
    "multiprocess >= 0.70.9",
    "scikit-learn >= 0.22.0",
    "xgboost >= 1.4",
    "tqdm >= 4",
    "drawsvg >= 2.0.0",
    "pyrodigal-gv >= 0.3",
    "diced >= 0.1.1",
    "pyhmmsearch >= 2025.10.23.post1",
    "pyhmmer >= 0.10.14, <0.11",
    "zstandard",
    "setuptools",
]

[project.urls]
Homepage = "https://github.com/Russel88/CRISPRCasTyper"
Download = "https://github.com/Russel88/CRISPRCasTyper/archive/v1.8.0.tar.gz"

[tool.setuptools]
include-package-data = true
script-files = ["bin/cctyper", "bin/repeatType", "bin/repeatTrain"]

[tool.setuptools.packages.find]
include = ["cctyper*"]

[tool.setuptools.package-data]
cctyper = ["data/**"]
